Moved PassportOrIdCardController to add directory. Changed amend-only route to follow change- pattern.
