+++ app/utils/print/IndividualPrintHelper.scala
@@ -41,7 +41,7 @@ class IndividualPrintHelper @Inject()(answerRowConverter: AnswerRowConverter,
       bound.dateQuestion(DateOfBirthPage, "individual.dateOfBirth", rts.DateOfBirthController.onPageLoad(NormalMode).url),
       bound.yesNoQuestion(NationalInsuranceNumberYesNoPage, "individual.nationalInsuranceNumberYesNo", rts.NationalInsuranceNumberYesNoController.onPageLoad(NormalMode).url),
       bound.ninoQuestion(NationalInsuranceNumberPage, "individual.nationalInsuranceNumber", rts.NationalInsuranceNumberController.onPageLoad(NormalMode).url),
-      bound.enumQuestion(PassportOrIdCardPage, "individual.passportOrIdCard", rts.PassportOrIdCardController.onPageLoad(NormalMode).url),
+      bound.enumQuestion(PassportOrIdCardPage, "individual.passportOrIdCard", addRts.PassportOrIdCardController.onPageLoad().url),
       bound.passportDetailsQuestion(PassportDetailsPage, "individual.passportDetails", addRts.PassportDetailsController.onPageLoad().url),
       bound.idCardDetailsQuestion(IdCardDetailsPage, "individual.idCardDetails", addRts.IdCardDetailsController.onPageLoad().url),
       bound.yesNoQuestion(LiveInTheUkYesNoPage, "individual.liveInTheUkYesNo", rts.LiveInTheUkYesNoController.onPageLoad(NormalMode).url),
