+++ test/utils/print/IndividualPrintHelperSpec.scala
@@ -104,7 +104,7 @@ class IndividualPrintHelperSpec extends SpecBase {
             AnswerRow(label = Html(messages("individual.name.checkYourAnswersLabel")), answer = Html("John Doe"), changeUrl = rts.NameController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.dateOfBirth.checkYourAnswersLabel", displayName)), answer = Html("9 March 1996"), changeUrl = rts.DateOfBirthController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.nationalInsuranceNumberYesNo.checkYourAnswersLabel", displayName)), answer = Html("No"), changeUrl = rts.NationalInsuranceNumberYesNoController.onPageLoad(mode).url),
-            AnswerRow(label = Html(messages("individual.passportOrIdCard.checkYourAnswersLabel", displayName)), answer = Html("Passport"), changeUrl = rts.PassportOrIdCardController.onPageLoad(mode).url),
+            AnswerRow(label = Html(messages("individual.passportOrIdCard.checkYourAnswersLabel", displayName)), answer = Html("Passport"), changeUrl = addRts.PassportOrIdCardController.onPageLoad().url),
             AnswerRow(label = Html(messages("individual.passportDetails.checkYourAnswersLabel", displayName)), answer = Html("France<br />123<br />3 February 2022"), changeUrl = addRts.PassportDetailsController.onPageLoad().url),
             AnswerRow(label = Html(messages("individual.liveInTheUkYesNo.checkYourAnswersLabel", displayName)), answer = Html("Yes"), changeUrl = rts.LiveInTheUkYesNoController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.ukAddress.checkYourAnswersLabel", displayName)), answer = Html("value 1<br />value 2<br />AB1 1AB"), changeUrl = rts.UkAddressController.onPageLoad(mode).url),
@@ -132,7 +132,7 @@ class IndividualPrintHelperSpec extends SpecBase {
             AnswerRow(label = Html(messages("individual.name.checkYourAnswersLabel")), answer = Html("John Doe"), changeUrl = rts.NameController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.dateOfBirth.checkYourAnswersLabel", displayName)), answer = Html("9 March 1996"), changeUrl = rts.DateOfBirthController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.nationalInsuranceNumberYesNo.checkYourAnswersLabel", displayName)), answer = Html("No"), changeUrl = rts.NationalInsuranceNumberYesNoController.onPageLoad(mode).url),
-            AnswerRow(label = Html(messages("individual.passportOrIdCard.checkYourAnswersLabel", displayName)), answer = Html("ID Card"), changeUrl = rts.PassportOrIdCardController.onPageLoad(mode).url),
+            AnswerRow(label = Html(messages("individual.passportOrIdCard.checkYourAnswersLabel", displayName)), answer = Html("ID Card"), changeUrl = addRts.PassportOrIdCardController.onPageLoad().url),
             AnswerRow(label = Html(messages("individual.idCardDetails.checkYourAnswersLabel", displayName)), answer = Html("France<br />123<br />3 February 2022"), changeUrl = addRts.IdCardDetailsController.onPageLoad().url),
             AnswerRow(label = Html(messages("individual.liveInTheUkYesNo.checkYourAnswersLabel", displayName)), answer = Html("No"), changeUrl = rts.LiveInTheUkYesNoController.onPageLoad(mode).url),
             AnswerRow(label = Html(messages("individual.nonUkAddress.checkYourAnswersLabel", displayName)), answer = Html("value 1<br />value 2<br />France"), changeUrl = rts.NonUkAddressController.onPageLoad(mode).url),
